--- tests/_task_models/dinov2_linear_semantic_segmentation/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (c) Lightly AG and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-#
